# Faster parsing sources to extract translation + add missing sources