```sh
[atlas]: ~/dev/taf/webapp/proton-vpn-settings [develop] 
$ cat po/template.pot| wc -l
4656
[atlas]: ~/dev/taf/webapp/proton-vpn-settings [develop] 
$ cat po/template.pot| wc -l
5442
``` 

Before:
```sh
/node_modules/proton-i18n/scripts/extract.sh po/template.pot  35,20s user 0,90s system 133% cpu 27,073 total
``` 

After
```sh
./node_modules/proton-i18n/scripts/extract.sh po/template.pot  21,15s user 0,62s system 139% cpu 15,597 total 
```